# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exchange API",
    "description": "API for the Exchange",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/admin": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Admin endpoint for test/dev operations",
        "description": "POST /api/admin\n\nHandles administrative operations like creating tokens, markets, and funding accounts.\nIn production, this endpoint should be protected or disabled.",
        "operationId": "admin_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Admin operation successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminRequest": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "ticker",
              "decimals",
              "name",
              "type"
            ],
            "properties": {
              "decimals": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              "name": {
                "type": "string"
              },
              "ticker": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "create_token"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "base_ticker",
              "quote_ticker",
              "tick_size",
              "lot_size",
              "min_size",
              "maker_fee_bps",
              "taker_fee_bps",
              "type"
            ],
            "properties": {
              "base_ticker": {
                "type": "string"
              },
              "lot_size": {
                "type": "string"
              },
              "maker_fee_bps": {
                "type": "integer",
                "format": "int32"
              },
              "min_size": {
                "type": "string"
              },
              "quote_ticker": {
                "type": "string"
              },
              "taker_fee_bps": {
                "type": "integer",
                "format": "int32"
              },
              "tick_size": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "create_market"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "user_address",
              "token_ticker",
              "amount",
              "signature",
              "type"
            ],
            "properties": {
              "amount": {
                "type": "string"
              },
              "signature": {
                "type": "string"
              },
              "token_ticker": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "faucet"
                ]
              },
              "user_address": {
                "type": "string"
              }
            }
          }
        ],
        "description": "Admin request with type discriminator"
      },
      "AdminResponse": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "token",
              "type"
            ],
            "properties": {
              "token": {
                "$ref": "#/components/schemas/Token"
              },
              "type": {
                "type": "string",
                "enum": [
                  "create_token"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "market",
              "type"
            ],
            "properties": {
              "market": {
                "$ref": "#/components/schemas/ApiMarket"
              },
              "type": {
                "type": "string",
                "enum": [
                  "create_market"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "user_address",
              "token_ticker",
              "amount",
              "new_balance",
              "type"
            ],
            "properties": {
              "amount": {
                "type": "string"
              },
              "new_balance": {
                "type": "string"
              },
              "token_ticker": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "faucet"
                ]
              },
              "user_address": {
                "type": "string"
              }
            }
          }
        ],
        "description": "Admin response with type discriminator"
      },
...
      "ApiMarket": {
        "type": "object",
        "description": "API representation of Market with String fields for JSON compatibility",
        "required": [
          "id",
          "base_ticker",
          "quote_ticker",
          "tick_size",
          "lot_size",
          "min_size",
          "maker_fee_bps",
          "taker_fee_bps"
        ],
        "properties": {
          "base_ticker": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lot_size": {
            "type": "string"
          },
          "maker_fee_bps": {
            "type": "integer",
            "format": "int32"
          },
          "min_size": {
            "type": "string"
          },
          "quote_ticker": {
            "type": "string"
          },
          "taker_fee_bps": {
            "type": "integer",
            "format": "int32"
          },
          "tick_size": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "Token": {
        "type": "object",
        "required": [
          "ticker",
          "decimals",
          "name"
        ],
        "properties": {
          "decimals": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "ticker": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}